[PyTorch] Move FP16 dot and GEMV kernels to new file in ATen/native/cpu/
Pull Request resolved: https://github.com/pytorch/pytorch/pull/137914

This is in preparation for supporting x86 as well; we need to
be in this directory so that we can get rebuilt with different
CPU_CAPABILITY settings (AVX2/AVX-512). Also incidentally starts
fulfilling request from @malfet to split the ARM64 fast path stuff
into its own file. BFloat16 will be in a later diff.
ghstack-source-id: 249885012

Differential Revision: [D64265755](https://our.internmc.facebook.com/intern/diff/D64265755/)