WIP Update schedules to use BACKWARD_INPUT / FULL_BACKWARD
Also, update tests to use B, I consistently.

Consider replacing '_validate_pipeline_order' with
'_simulate_comms_compute' since they are somewhat duplicate
functionality, and it is also annoying to update 'validate' to handle
seprate BACKWARD_INPUT.  Validate is different in that it considers the
compute-only schedule, while simulate considers the compute+comms
schedule.

However, simulate is too slow for some of the unit tests with extra
large numbers of ranks and microbatches, and is also failing to execute
some of the schedules.  After fixing the failures, I will see if it is
easy to optimize the simulator or not.

- Optimized simulate and add_send_recv, tests are fast now even at large
  scales
- need to run other tests and do cleanup work, then this should be ready
  to land

ghstack-source-id: eb263f206f5ea547a316a535d2f8a1e277bfd894
Pull Request resolved: https://github.com/pytorch/pytorch/pull/138886